+++ packages/eslint-plugin/src/rules/explicit-module-boundary-types.ts
@@ -427,7 +427,11 @@ export default util.createRule<Options, MessageIds>({
       const isConstructor =
         node.parent?.type === AST_NODE_TYPES.MethodDefinition &&
         node.parent.kind === 'constructor';
-      if (!isConstructor && !node.returnType) {
+      const isSetAccessor =
+        (node.parent?.type === AST_NODE_TYPES.TSAbstractMethodDefinition ||
+          node.parent?.type === AST_NODE_TYPES.MethodDefinition) &&
+        node.parent.kind === 'set';
+      if (!isConstructor && !isSetAccessor && !node.returnType) {
         context.report({
           node,
           messageId: 'missingReturnType',