+++ packages/eslint-plugin/src/rules/no-implied-eval.ts
@@ -35,6 +35,7 @@ export default util.createRule({
   defaultOptions: [],
   create(context) {
     const parserServices = util.getParserServices(context);
+    const program = parserServices.program;
     const checker = parserServices.program.getTypeChecker();
 
     function getCalleeName(
@@ -113,14 +114,29 @@ export default util.createRule({
     function checkImpliedEval(
       node: TSESTree.NewExpression | TSESTree.CallExpression,
     ): void {
+      const tsNode = parserServices.esTreeNodeToTSNodeMap.get(node.callee);
+      const type = checker.getTypeAtLocation(tsNode);
+
       const calleeName = getCalleeName(node.callee);
       if (calleeName === null) {
         return;
       }
 
       if (calleeName === FUNCTION_CONSTRUCTOR) {
-        context.report({ node, messageId: 'noFunctionConstructor' });
-        return;
+        const symbol = type.getSymbol();
+        if (symbol) {
+          const declarations = symbol.getDeclarations() ?? [];
+          for (const declaration of declarations) {
+            const sourceFile = declaration.getSourceFile();
+            if (program.isSourceFileDefaultLibrary(sourceFile)) {
+              context.report({ node, messageId: 'noFunctionConstructor' });
+              return;
+            }
+          }
+        } else {
+          context.report({ node, messageId: 'noFunctionConstructor' });
+          return;
+        }
       }
 
       if (node.arguments.length === 0) {