+++ packages/eslint-plugin/src/rules/no-unnecessary-condition.ts
@@ -165,7 +165,18 @@ export default createRule<Options, MessageId>({
      * Checks if a conditional node is necessary:
      * if the type of the node is always true or always false, it's not necessary.
      */
-    function checkNode(node: TSESTree.Expression): void {
+    function checkNode(
+      node: TSESTree.Expression,
+      isUnaryNotArgument = false,
+    ): void {
+      // Check if the node is Unary Negation expression and handle it
+      if (
+        node.type === AST_NODE_TYPES.UnaryExpression &&
+        node.operator === '!'
+      ) {
+        return checkNode(node.argument, true);
+      }
+
       // Since typescript array index signature types don't represent the
       //  possibility of out-of-bounds access, if we're indexing into an array
       //  just skip the check, to avoid false positives
@@ -175,7 +186,14 @@ export default createRule<Options, MessageId>({
 
       // When checking logical expressions, only check the right side
       //  as the left side has been checked by checkLogicalExpressionForUnnecessaryConditionals
-      if (node.type === AST_NODE_TYPES.LogicalExpression) {
+      //
+      // Unless the node is nullish coalescing, as it's common to use patterns like `nullBool ?? true` to to strict
+      //  boolean checks if we inspect the right here, it'll usually be a constant condition on purpose.
+      // In this case it's better to inspect the type of the expression as a whole.
+      if (
+        node.type === AST_NODE_TYPES.LogicalExpression &&
+        node.operator !== '??'
+      ) {
         return checkNode(node.right);
       }
 
@@ -193,13 +211,15 @@ export default createRule<Options, MessageId>({
       ) {
         return;
       }
-      const messageId = isTypeFlagSet(type, ts.TypeFlags.Never)
-        ? 'never'
-        : !isPossiblyTruthy(type)
-        ? 'alwaysFalsy'
-        : !isPossiblyFalsy(type)
-        ? 'alwaysTruthy'
-        : undefined;
+      let messageId: MessageId | null = null;
+
+      if (isTypeFlagSet(type, ts.TypeFlags.Never)) {
+        messageId = 'never';
+      } else if (!isPossiblyTruthy(type)) {
+        messageId = !isUnaryNotArgument ? 'alwaysFalsy' : 'alwaysTruthy';
+      } else if (!isPossiblyFalsy(type)) {
+        messageId = !isUnaryNotArgument ? 'alwaysTruthy' : 'alwaysFalsy';
+      }
 
       if (messageId) {
         context.report({ node, messageId });
@@ -218,13 +238,15 @@ export default createRule<Options, MessageId>({
       if (isTypeAnyType(type) || isTypeUnknownType(type)) {
         return;
       }
-      const messageId = isTypeFlagSet(type, ts.TypeFlags.Never)
-        ? 'never'
-        : !isPossiblyNullish(type)
-        ? 'neverNullish'
-        : isAlwaysNullish(type)
-        ? 'alwaysNullish'
-        : undefined;
+
+      let messageId: MessageId | null = null;
+      if (isTypeFlagSet(type, ts.TypeFlags.Never)) {
+        messageId = 'never';
+      } else if (!isPossiblyNullish(type)) {
+        messageId = 'neverNullish';
+      } else if (isAlwaysNullish(type)) {
+        messageId = 'alwaysNullish';
+      }
 
       if (messageId) {
         context.report({ node, messageId });