+++ packages/eslint-plugin/src/rules/no-unsafe-assignment.ts
@@ -30,7 +30,7 @@ export default util.createRule({
       unsafeArrayPatternFromTuple:
         'Unsafe array destructuring of a tuple element with an any value.',
       unsafeAssignment:
-        'Unsafe asignment of type {{sender}} to a variable of type {{receiver}}.',
+        'Unsafe assignment of type {{sender}} to a variable of type {{receiver}}.',
       unsafeArraySpread: 'Unsafe spread of an any value in an array.',
     },
     schema: [],
@@ -238,6 +238,11 @@ export default util.createRule({
       );
 
       if (util.isTypeAnyType(senderType)) {
+        // handle cases when we assign any ==> unknown.
+        if (util.isTypeUnknownType(receiverType)) {
+          return false;
+        }
+
         context.report({
           node: reportingNode,
           messageId: 'anyAssignment',