+++ packages/scope-manager/src/lib/es2020.ts
@@ -8,6 +8,7 @@ import { es2020_bigint } from './es2020.bigint';
 import { es2020_promise } from './es2020.promise';
 import { es2020_string } from './es2020.string';
 import { es2020_symbol_wellknown } from './es2020.symbol.wellknown';
+import { es2020_intl } from './es2020.intl';
 
 export const es2020 = {
   ...es2019,
@@ -15,4 +16,5 @@ export const es2020 = {
   ...es2020_promise,
   ...es2020_string,
   ...es2020_symbol_wellknown,
+  ...es2020_intl,
 } as Record<string, ImplicitLibVariableOptions>;