+++ packages/scope-manager/tests/util/serializers/Reference.ts
@@ -4,8 +4,10 @@ import { Reference } from '../../../src/referencer/Reference';
 const serializer = createSerializer(Reference, [
   'identifier',
   'init',
+  'isRead',
   'isTypeReference',
   'isValueReference',
+  'isWrite',
   'resolved',
   'writeExpr',
 ]);