+++ packages/shared-fixtures/fixtures/typescript/types/tuple-named-type.src.ts
@@ -1,0 +1,1 @@
+type Foo = [a: string, b?: string]