+++ packages/shared-fixtures/fixtures/typescript/types/tuple-named.src.ts
@@ -1,0 +1,1 @@
+let x: [a: number, b: number, c: number];